add form's overflow animation feature support

+++ frameworks/js/napi/form_provider/js_form_provider.h
@@ -100,7 +100,7 @@ private:
     napi_value OnDeactivateSceneAnimation(napi_env env, size_t argc, napi_value* argv);
     static bool ConvertFormOverflowInfo(napi_env env, napi_value argv, AppExecFwk::OverflowInfo* overflowInfo);
     static bool ConvertOverflowInfoArea(napi_env env, napi_value rangeArea, AppExecFwk::Rect &area);
-    static bool GetAndConvertProperty(napi_env env, napi_value object, const char* propertyName, int32_t& outValue);
+    static bool GetAndConvertProperty(napi_env env, napi_value object, const char* propertyName, double& outValue);
 };
 }  // namespace AbilityRuntime
 }  // namespace OHOS
